UI and Sailing

+++ SailingControl.cpp
@@ -44,8 +44,8 @@ bool SailingControl::createSailing(const string& sailingId, const string& vessel
         return false;
     }
 
-    double hcll = vessel.getHCLL(vesselName);
-    double lcll = vessel.getLCLL(vesselName);
+    int hcll = vessel.getHCLL(vesselName);
+    int lcll = vessel.getLCLL(vesselName);
 
     return Sailing::writeSailing(const_cast<string&>(sailingId),
                                  const_cast<string&>(vesselName),
